#38 Make the channel ID a String
It is a String in JDA, too. This way, the conversion
to Long gets obsolete.

+++ app/src/main/kotlin/de/dagadeta/schlauerbot/wordchaingame/DiscordWordChainGame.kt
@@ -68,7 +68,7 @@ class DiscordWordChainGame(
     }
 
     override fun onMessageReceived(event: MessageReceivedEvent) {
-        if (event.channel.id.toLong() != wordChainGameConfig.channelId || event.author.isBot) return
+        if (event.channel.id != wordChainGameConfig.channelId || event.author.isBot) return
         game.onMessageReceived(event.author.id, event.message.contentDisplay)
             .onFailure { answer -> sendInvalidWordMessage(event.message, answer) }
     }
